Allmost Final version before release

+++ src/app/components/commons/delete-button/delete-button.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ArticleService } from 'src/app/services/ArticleService/article.service';
 import { AuthService } from 'src/app/services/AuthService/auth.service';
+import { ModalService } from 'src/app/services/ModalService/modal.service';
 
 @Component({
   selector: 'app-delete-button',
@@ -12,7 +13,8 @@ export class DeleteButtonComponent implements OnInit {
   @Input('slugArticle') slugArticle
   constructor(private authService: AuthService,
               private router: Router,
-              private articleService: ArticleService) { }
+              private articleService: ArticleService,
+              public dialog: ModalService) { }
 
   ngOnInit(): void {
   }
@@ -23,9 +25,12 @@ export class DeleteButtonComponent implements OnInit {
       this.router.navigate(['/', 'login']);
       return;
     }
-    this.articleService.deleteArticle(this.slugArticle.slug).subscribe((res) => {
-      this.router.navigate(['/']);
-    });
-  }
-
+    this.dialog.openConfirmDialog().afterClosed().subscribe(res => {
+      if (res) {
+        this.articleService.deleteArticle(this.slugArticle.slug).subscribe((res) => {
+          this.router.navigate(['profile']);
+          });
+        }   
+      })
+    } 
 }
